Validation logic added to: 1. Search/Insert/Delete users

+++ Users/Services/UsersService.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using WorkManagement.Models;
+using WorkManagement.Models.Shared;
 using WorkManagement.Users.BL;
 using WorkManagement.Users.Models;
 
@@ -62,8 +63,15 @@ namespace WorkManagement.Users.Services
         {
             var response = new ResponseViewModel<bool>();
 
-            response.Data = m_userBL.DeleteUser(userId);
-            response.ErrorDesc = response.Data ? string.Empty : "מחיקת משתמש לא הצליחה.";
+            if (userId != SessionHelper.SessionUser.UserId)
+            {
+                response.Data = m_userBL.DeleteUser(userId);
+                response.ErrorDesc = response.Data ? string.Empty : "מחיקת משתמש לא הצליחה.";
+            }
+            else
+            {
+                response.ErrorDesc = response.Data ? string.Empty : "אי אפשר למחוק את עמצך.";
+            }
 
             return response;
         }
@@ -96,7 +104,7 @@ namespace WorkManagement.Users.Services
                     response.ErrorDesc = "לא נמאו משתמשים לפי נתוני חיפוש.";
                 }
             }
-            catch(Exception e)
+            catch(Exception)
             {
                 response.ErrorDesc = "קרתה תקלה בחיפוש משתמשים.";
             }
